requirements.txt
google-api-python-client
google-auth-oauthlib
langchain
langchain-community
langchain-google-genai
python-dotenv
requests
